extract chatmessage to baseprovider
