+++ src/main/presenter/llmProviderPresenter/providers/siliconcloudProvider.ts
@@ -1,6 +1,7 @@
 import { LLM_PROVIDER, LLMResponse, LLMResponseStream } from '@shared/presenter'
-import { ChatMessage, OpenAICompatibleProvider } from './openAICompatibleProvider'
+import { OpenAICompatibleProvider } from './openAICompatibleProvider'
 import { ConfigPresenter } from '../../configPresenter'
+import { ChatMessage } from '../baseProvider'
 
 export class SiliconcloudProvider extends OpenAICompatibleProvider {
   constructor(provider: LLM_PROVIDER, configPresenter: ConfigPresenter) {
@@ -8,7 +9,7 @@ export class SiliconcloudProvider extends OpenAICompatibleProvider {
   }
 
   async completions(
-    messages: { role: 'system' | 'user' | 'assistant'; content: string }[],
+    messages: ChatMessage[],
     modelId: string,
     temperature?: number,
     maxTokens?: number
